requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pyyaml>=5.4.1
requests>=2.26.0
beautifulsoup4>=4.9.3
selenium>=4.1.0
webdriver-manager>=3.5.2

# Social Media APIs
tweepy>=4.4.0
facebook-sdk>=3.1.0
python-linkedin-v2>=0.8.14
instaloader>=4.9.5

# Web Scraping
scrapy>=2.5.0
playwright>=1.33.0

# Data Processing
pandas>=1.3.3
numpy>=1.21.2

# Logging and Monitoring
loguru>=0.5.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1